// 3_HEXAGONAL_ARCH/msservicechangeiot/src/share/config/dtos/data-config-infraestructure.dto.ts
export class DataConfigInfraestructureDto{
  application_context: string;
  application_name: string;
  operation_connectivity: string;
  timeout: number;
  database_user_1: string;
  database_pass_1: string;
  database_string_1: string;
  database_procedure_path_1: string;
  database_user_2: string;
  database_pass_2: string;
  database_string_2: string;
  database_procedure_path_2: string;
}